requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
dj-rest-auth==2.2.4
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
